tg_bot/requirements.txt
aiogram==2.10.1
attrs
asyncpg
